Move signing logic to its own module
